chore: move makeConfig to utils
