+++ packages/core/src/config/config.ts
@@ -21,7 +21,6 @@ import { ProblemSeverity, NormalizedProblem } from '../walk';
 
 import recommended from './recommended';
 import { NodeType } from '../types';
-import { defaultPlugin } from './builtIn';
 
 export const IGNORE_FILE = '.redocly.lint-ignore.yaml';
 const IGNORE_BANNER =
@@ -623,13 +622,4 @@ function assignExisting<T>(target: Record<string, T>, obj: Record<string, T>) {
       target[k] = obj[k];
     }
   }
-}
-
-export function makeConfig(rules: Record<string, RuleConfig>, decorators?:  Record<string, DecoratorConfig>) {
-  return new LintConfig({
-    plugins: [defaultPlugin],
-    extends: [],
-    rules,
-    decorators,
-  });
 }
