+++ packages/core/src/rules/__tests__/no-unresolved-refs.test.ts
@@ -2,8 +2,7 @@ import path = require('path');
 import { outdent } from 'outdent';
 import { lintDocument } from '../../lint';
 import { BaseResolver } from '../../resolve';
-import { parseYamlToDocument, replaceSourceWithRef } from '../../../__tests__/utils';
-import { makeConfig } from '../../config/config';
+import { parseYamlToDocument, replaceSourceWithRef, makeConfig } from '../../../__tests__/utils';
 
 describe('oas3 boolean-parameter-prefixes', () => {
   it('should report on unresolved $ref', async () => {
