+++ packages/core/src/rules/common/__tests__/info-description.test.ts
@@ -1,9 +1,7 @@
 import { outdent } from 'outdent';
 import { lintDocument } from '../../../lint';
-import { parseYamlToDocument, replaceSourceWithRef } from '../../../../__tests__/utils';
+import { parseYamlToDocument, replaceSourceWithRef, makeConfig } from '../../../../__tests__/utils';
 import { BaseResolver } from '../../../resolve';
-
-import { makeConfig } from '../../../config/config';
 
 describe('Oas3 info-description', () => {
   it('should report on info with no description', async () => {
