+++ packages/core/src/rules/oas2/__tests__/spec/referenceableScalars.test.ts
@@ -1,8 +1,7 @@
 import { outdent } from 'outdent';
-import { parseYamlToDocument, replaceSourceWithRef } from '../../../../../__tests__/utils';
+import { parseYamlToDocument, replaceSourceWithRef, makeConfig } from '../../../../../__tests__/utils';
 import { lintDocument } from '../../../../lint';
 import { BaseResolver } from '../../../../resolve';
-import { makeConfig } from '../../../../config/config';
 
 describe('Referenceable scalars', () => {
   it('should not report $ref description', async () => {
