+++ packages/core/src/rules/oas3/__tests__/boolean-parameter-prefixes.test.ts
@@ -1,6 +1,5 @@
 import { outdent } from 'outdent';
-import { parseYamlToDocument, replaceSourceWithRef } from '../../../../__tests__/utils';
-import { makeConfig } from '../../../config/config';
+import { parseYamlToDocument, replaceSourceWithRef, makeConfig } from '../../../../__tests__/utils';
 import { lintDocument } from '../../../lint';
 import { BaseResolver } from '../../../resolve';
